gtp_packet: Export {enc_dec}ode_protocol_config_opts
gtp_packet message enc/dec works with decoded PCO. Whoever, sometimes
the user of the message has an already encoded PCO which needs to be
forwarded. It can also happen that the user needs to simply forward the
received PCO in binary form.
Export the functions to enc/dec the IE, so that the user can also handle
binary PCO data.